login 2

+++ studybud/base/views.py
@@ -3,11 +3,18 @@ from django.contrib import messages
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -20,10 +27,20 @@ def loginPage(request):
         user = authenticate(request, username=username, password=password)
         if user is not None:
             login(request, user)
-            return redirect('home') 
+            return redirect('home')
+        else:
+            messages.error(request, 'User OR Password does not exist!')    
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+# def userRegistration(request):
+
+
 
 
 def home(request):
@@ -46,6 +63,7 @@ def room(request, pk):
     context = {'room': room}
     return render (request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -57,9 +75,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        #return HttpResponse('You are NOT') 
+        messages.warning(request, 'You are NOT allowed to update!')
+        return redirect('home')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -70,8 +94,16 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        #return HttpResponse('You are NOT') 
+        messages.error(request, 'You are NOT allowed to Delete!')
+        return redirect('home')
+
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
